convert a GRanges to a data.frame that can be written out as a POSTable file for OpenCRAVAT

convert a GRanges to a data.frame that can be written out as a POSTable file for OpenCRAVAT

make_oc_POSTable(gr, chain = NULL, genome = NULL)

Arguments

gr

GRanges instance

chain

rtracklayer chain for liftOver

genome

character(1) genome tag

Examples

tf = tempfile()
mtb = make_oc_POSTable(gr38)
write.table(mtb, file=tf, sep="\t", col.names=FALSE, row.names=FALSE, quote=FALSE)
read.delim(tf, sep="\t", nrow=5, h=FALSE)
#>     V1        V2 V3 V4                           V5  V6
#> 1 chr1  11501469  G  A TCGA-OR-A5J1-01A-11D-A29I-10 v_1
#> 2 chr1  12249327  T  G TCGA-OR-A5J1-01A-11D-A29I-10 v_2
#> 3 chr1  33354414  C  T TCGA-OR-A5J1-01A-11D-A29I-10 v_3
#> 4 chr1 152827646  C  T TCGA-OR-A5J1-01A-11D-A29I-10 v_4
#> 5 chr1 152827655  C  T TCGA-OR-A5J1-01A-11D-A29I-10 v_5