utility for working with multi-experiment results from ENCODE TFBS in opencravat

utility for working with multi-experiment results from ENCODE TFBS in opencravat

pullTF(x, tf = "YY1")

Arguments

x

a properly filtered data.frame with the relevant columns

tf

character(1) symbol for TF as recorded by ENCODE

Note

The issue is that many TF experiments are recorded in a single excel spreadsheet cell. A column gives the TFs studied with semicolon delimiter, and other columns give the cell line, quality of result. These columns should be coupled with relevant variant-level metadata

Examples

if (FALSE) { # \dontrun{
x = read.csv("IPF GWAS Variants - Variant.csv")
inds = grep("YY1", x$ENCODE.TFBS)
xyy = x[inds,]
dd = data.frame(t(apply(xyy[,19:23], 1, pullTF)))
hh = cbind(xyy, dd)
impactYY1 = hh[,c(1:10,43,54:58)]
write.csv(impactYY1, "encodeYY1.csv")
} # }